Configure import order with eslint
